+++ st2common/st2common/config.py
@@ -318,6 +318,13 @@ def register_opts(ignore_errors=False):
             default=None,
             help="Login method to use (AMQPLAIN, PLAIN, EXTERNAL, etc.).",
         ),
+        cfg.StrOpt(
+            "compression",
+            default=None,
+            help="Compression algorithm to use for compressing the payloads which are send over "
+            "the message bus. Valid values include: zstd, lzma, bz2, gzip. Defaults to no "
+            "compression.",
+        ),
     ]
 
     do_register_opts(messaging_opts, "messaging", ignore_errors)